// src/interfaces/UserInterface.ts
// Generated by https://quicktype.io

export interface User {
    userId:   string;
    userData: UserData;
}

export interface UserData {
    name:    string;
    mail:    string;
    phone:   string;
    surname: string;
}

export interface LogResponse {
    userName: string;
    userId: string;
}

export interface LoginData {
    mail: string;
    password: string
}

export interface EditUserResponse {
    message:     string;
    metadata:    Metadata;
    newUserData: NewUserData;
}

export interface Metadata {
    $metadata: MetadataClass;
}

export interface MetadataClass {
    httpStatusCode:  number;
    requestId:       string;
    attempts:        number;
    totalRetryDelay: number;
}

export interface NewUserData {
    name:    string;
    phone:   string;
    surname: string;
    mail:    string;
}